Fix/unsyncedchanges (#635) (#654)
This adds an improved way of figuring out if the document is fully synced with the server.

WARNING: You should update @hocuspocus/server before upgrading the client. "hasUnsyncedChanges" of @hocuspocus/server <2.3.0 is NOT compatible with @hocuspocus/provider >=2.3.0

If you upgrade the server while running provider <2.2.3, the provider will show errors in the browser, which can be safely ignored. If you dont want to see errors, upgrade the provider to v2.2.3 first, then upgrade the server to 2.3.0, and then upgrade the provider again. Or just upgrade everything simultaneously to 2.3.0.